+++ src/main/java/com/opencastsoftware/wasm4j/instructions/reference/RefFunc.java
@@ -1,13 +1,13 @@
 package com.opencastsoftware.wasm4j.instructions.reference;
 
 public class RefFunc implements ReferenceInstruction {
-    private final long funcIndex;
+    private final int funcIndex;
 
-    public RefFunc(long funcIndex) {
+    public RefFunc(int funcIndex) {
         this.funcIndex = funcIndex;
     }
 
-    public long funcIndex() {
+    public int funcIndex() {
         return funcIndex;
     }
 }
